Refactor the drive finding code into a separate file for testing.
To make it easier to locate the drives when we're testing, the
drive finding code has been moved to a separate file. This means
that we can test its behaviour with a simple test program which
steps through the drives and reports what discs it found.

At the moment that's all it does, but it's enough to exercise that
code and prove that it returns the right details. Later we can add
some tests for the FS specific code.

Tested on the tentative implementation in Pyromaniac and appears
to report the drives that exist. As the disc names are processed
by the add_drive function, there isn't currently anything special
to add here.