Fix for add_drive dereferencing zero-page. Update startup app.
The add_drive function wasn't checking a pointer before using it -
this meant that unnamed discs would access zero page. We now use the
name :# when creating such discs.

When the module was initialised it would start the task implicitly -
that's odd but not unreasonable - but it would do so with a non-X
SWI in SVC mode, which if it fails will call the application handler.
We now use the _swix call.