Update the logic for determining removeable drives.
The drives which are removeable and which are fixed are determined
by the set of flags supplied to FileCore when it was registered.
These flags can be read to determine whether the discs are actually
removeable or not.

Similarly the discs which are absent can be determined by the flags
in the ReadStatus call.

This changes the numbers of drives that are declared as being available
and whether they really exist or not, and means that we can now pass
through a flag to the callback function which indicates whether the
disc is removeable - and that can be used to report the type of disc
icon on the icon bar.

The code is a bit more complex, but I think it's doing the right
thing.