Add manipulation of the Tool commands.
The tool commands are able to locate variables which define the tools
for operating on discs in a generic way. This should allow us to
invoke them as separate commands so that they can be developed and
tested separately (and replaced by the user if necessary).

At present, only the initial functions have been created which give
the ability to determine what the command would be for the invocation
of the tool. Strictly that's all that's necessary for the functioning
of the feature - but it would be better if we also sent around the
messages as well.

Tests have been written to check the behaviour of the tool as well.
This only covers the variable at present.