Refactor the menu opening code to have an update function.
The menu state will change dynamically depending on the FS and
drive that are selected. This means that we'll need to update more
than just the title. The same code will be required for the
recreation of the menu when someone clicks Adjust, so we may
as well have it in a separate function.