Update the menu options to use constants; add invocation of tools.
The tools interface is now invoked when we call the menu items.
This should be greying the menu items when the option is not
available.

The menu structures have been updated so that they use symbol names
to make it easier to see which items are present and which are
not.