Move the Tool invocation handling into the FilerX code.
The FilerX code appears to be the thing that handles FS and Drive
operations. So moving the code in there means that it's easier
to operate on them in the place that they're handled.

The floppy flag is now used to indicate whether the device is
mounted/formatted as a floppy device.