Replace all the File system determination with an interface module.
We now have a module FSParameters which is intended to handle the
translation from the name you gave to the name that can be used for
SWI calls, and the name that is used for the filing system prefix -
which might be different things. We handle the very special case of
RAMFS which has a SWI prefix of RamFS.

This has meant replacing all the names in the system with references
through this structure. It's not perfect but it's now more reliable
than it was, and a RAM disc is addable as an icon.