Tool operations (#24)
* Make the use of the boolean into a common header.

It's tedious having to dupliate that header information. Move it to
a separate common file.

* Add manipulation of the Tool commands.

The tool commands are able to locate variables which define the tools
for operating on discs in a generic way. This should allow us to
invoke them as separate commands so that they can be developed and
tested separately (and replaced by the user if necessary).

At present, only the initial functions have been created which give
the ability to determine what the command would be for the invocation
of the tool. Strictly that's all that's necessary for the functioning
of the feature - but it would be better if we also sent around the
messages as well.

Tests have been written to check the behaviour of the tool as well.
This only covers the variable at present.

* Refactor the menu opening code to have an update function.

The menu state will change dynamically depending on the FS and
drive that are selected. This means that we'll need to update more
than just the title. The same code will be required for the
recreation of the menu when someone clicks Adjust, so we may
as well have it in a separate function.

* Update the menu options to use constants; add invocation of tools.

The tools interface is now invoked when we call the menu items.
This should be greying the menu items when the option is not
available.

The menu structures have been updated so that they use symbol names
to make it easier to see which items are present and which are
not.

* Pass through logical flags for ejectable and floppy-like discs.

The FindDrives code now decides whether the discs are ejectable or
floppy like. These are then stored in the Drive structure so that
we know what type of device they are.

* Add the Eject menu item to the menu, if supported.

The Eject menu item isn't usually on the iconbar. But we can add
it if we want - it's not currently supported though.

* Move the Tool invocation handling into the FilerX code.

The FilerX code appears to be the thing that handles FS and Drive
operations. So moving the code in there means that it's easier
to operate on them in the place that they're handled.

The floppy flag is now used to indicate whether the device is
mounted/formatted as a floppy device.

* Make the Eject menu item shade when not ejectable.

The Eject item should only be available when the device claims it
is ejectable.

* Add some special casing for reading RAMFS.

The RAMFS drives aren't returned returned properly, because it has
different names. This small change just special cases the name,
and allows the lack of the MiscOp SWI to not be fatal.

* Replace all the File system determination with an interface module.

We now have a module FSParameters which is intended to handle the
translation from the name you gave to the name that can be used for
SWI calls, and the name that is used for the filing system prefix -
which might be different things. We handle the very special case of
RAMFS which has a SWI prefix of RamFS.

This has meant replacing all the names in the system with references
through this structure. It's not perfect but it's now more reliable
than it was, and a RAM disc is addable as an icon.