Add support for default filing system icons.
Because it's useful for some filing systems to have different icons,
we now allow a table of built in icons for the filing systems we
know should be differentiated. Primarily this is to allow RAMFS to
have a different icon to ADFS. Otherwise you end up with two disc
drive icons beside one another, which looks unhelpful.