Move the 'next' pointer in the FS structure to the start.
It is generally the case that the structure chaining pointers be at the
start of the structure. The lack of this meant that we weren't writing
to low addresses in memory, and so didn't trigger a warning in the
CI.

So really this is just a change to highlight a problem in the code.