Move the FS default sprites database to a separate file.
The FS default sprites database had been part of the Icon code, but
it's not really part of the Icon system. It's a standalone database
of details about the filing systems. It should be separate. It can
then be tested separately to make sure that it does what it needs
to.

This change moves the sprites database to a separate file, and creates
the FSNumbers header which can be used in a more general way. The
FSNumbers header mirrors the C:riscos.h.FSNumbers header, which is
present in my build environment but not elsewhere.

Also fixes a couple of the makefiles that were failing to run
properly.