Wire in the icon priorities to the iconbar creation.
The icon priorities should now be present in the creation of the
iconbar icons. I'm not convinced that the code is correct at the
moment, as I think we only set the priority for the first filing
system that is created. However, this at least gives us a starting
point for the priorities to be used.