Add support for default filing system sprites and priorities. (#29)
* Add support for default filing system icons.

Because it's useful for some filing systems to have different icons,
we now allow a table of built in icons for the filing systems we
know should be differentiated. Primarily this is to allow RAMFS to
have a different icon to ADFS. Otherwise you end up with two disc
drive icons beside one another, which looks unhelpful.

* Fix for tests failing with newly named sprites.

The new sprite names weren't expected by the test code, so the tests
failed. The tests have been updated to reflect the RAMFS filing system
having a different icon.

* Move the FS default sprites database to a separate file.

The FS default sprites database had been part of the Icon code, but
it's not really part of the Icon system. It's a standalone database
of details about the filing systems. It should be separate. It can
then be tested separately to make sure that it does what it needs
to.

This change moves the sprites database to a separate file, and creates
the FSNumbers header which can be used in a more general way. The
FSNumbers header mirrors the C:riscos.h.FSNumbers header, which is
present in my build environment but not elsewhere.

Also fixes a couple of the makefiles that were failing to run
properly.

* Add support for priority returned by the FSDefaults code.

We now keep a database of the default priority to use for each filing
system. This should be able to be used to find the right place to put
the new icons when they are added.

* Wire in the icon priorities to the iconbar creation.

The icon priorities should now be present in the creation of the
iconbar icons. I'm not convinced that the code is correct at the
moment, as I think we only set the priority for the first filing
system that is created. However, this at least gives us a starting
point for the priorities to be used.

* Update all the test code to use the FSNumbers header.

Using the header makes it easier to reason about things and make sure
that what we typed is actually the number we meant.

* Update the bar icon creation algorithm to use priority lists.

We're essentially treating each priority on the iconbar as its own
list of icons, but keeping within the existing scheme of adding the
icons relative to the order they were added.