Update the CMHG file to use VersionNum to track versions.
Rather than having the version number hard coded into the files, and
to make it easier to update the version numbering, the CMHG file can
include a file which contains this information. This file can be
managed by a number of tools that have been created over the years.

Also corrects a couple of typos in the help messages.